ajout middleware pour admin

+++ backend/controllers/post.js
@@ -15,9 +15,11 @@ const createOnePost = async (req, res, next) => {
     const tokenId = decodedToken.id
     // console.log("TOKEN DANS CREATE POST", tokenId);
     console.log(req.file);
-    await Post.create({ ...body,
-                        urlImage: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-                        userId: tokenId })
+    await Post.create({
+        ...body,
+        urlImage: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+        userId: tokenId
+    })
         .then(() => { res.status(201).json({ message: "Votre post à été publié !" }) })
         .catch(error => res.status(500).json({ error: error }))
 }
@@ -34,7 +36,7 @@ const getOnePost = async (req, res, next) => {
     try {
         // Recherche de l'utilisateur et vérification
         let post = await Post.findOne({ where: { id: id }, include: Comment, raw: true })
-        
+
         if (post === null) {
             return res.status(404).json({ message: 'Ce post n\'existe pas !' })
         }
@@ -43,39 +45,49 @@ const getOnePost = async (req, res, next) => {
     } catch {
         return res.status(500).json({ message: 'Erreur de base de donnée du getOne' })
     }
-    
+
 }
 
 const updatePost = async (req, res, next) => {
     const { id } = req.params
-    if(req.file) {
-        await Post.findOne({ where: { id: id}, raw: true})
+    if (req.file) {
+        await Post.findOne({ where: { id: id }, raw: true })
             .then(post => {
+                // VERIFICATION DE L'ID POUR NE POUVOIR MODIFIER QUE SON PROPRE POST
+                if (post.userId !== req.body.userId) {
+                    return res.status(403).json({ message: 'Vous n\'avez pas le droit de faire celà' })
+                }
                 const filename = post.urlImage.split('/images/')[1];
                 fs.unlink(`images/${filename}`, async () => {
-                    const postUpdated = 
-                    {   
+                    const postUpdated =
+                    {
                         ...req.body,
                         urlImage: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
                     }
                     console.log(postUpdated);
-                    await Post.update(post = postUpdated, { where: { id : id}})
-                    return res.status(201).json({message: 'Photo de post modifiée.'})
-            })
-            }).catch(error => res.status(400).json({ message : "error" }));
+                    await Post.update(post = postUpdated, { where: { id: id } })
+                    return res.status(201).json({ message: 'Photo de post modifiée.' })
+                })
+            }).catch(error => res.status(400).json({ message: "error" }));
     }
     else {
-        try{
+        try {
             // Recherche du post et vérification non nul
-            let post = await Post.findOne({ where: {id: id}, raw: true})
-            if(post === null){
-                return res.status(404).json({ message: 'Ce post n\'existe pas !'})
+            // ON NE TROUVE QUE LES POST QUE L'ON A CREE NOUS MEME
+            let post = await Post.findOne({ where: { id: id/*, userId : req.body.userId*/ }, raw: true })
+            console.log(post);
+            // VERIFICATION DE L'ID POUR NE POUVOIR MODIFIER QUE SON PROPRE POST
+            if (post.userId !== req.body.userId) {
+                return res.status(403).json({ message: 'Vous n\'avez pas le droit de faire celà' })
+            }
+            if (post === null) {
+                return res.status(404).json({ message: 'Ce post n\'existe pas !' })
             }
             // Mise à jour de l'utilisateur
-            await Post.update( req.body, { where: {id: id}})
-            return res.json({ message: 'post modifié avec succès !'})
-        }catch(error){
-            return res.status(500).json({ message: 'Erreur de base de donnée', error: error })
+            await Post.update(req.body, { where: { id: id } })
+            return res.json({ message: 'post modifié avec succès !' })
+        } catch (error) {
+            return res.status(500).json({ message: 'Erreur de base de donnée', error })
         }
     }
 }
@@ -83,22 +95,27 @@ const updatePost = async (req, res, next) => {
 const deletePost = async (req, res, next) => {
     const { id } = req.params;
     try {
-        const post = await Post.findOne({ where: { id : id }, raw: true})
+        const post = await Post.findOne({ where: { id: id }, raw: true })
+        // VERIFICATION DE L'ID POUR NE POUVOIR MODIFIER QUE SON PROPRE POST
+        console.log(post);
+        // if (post.userId !== req.body.userId) {
+        //     return res.status(403).json({ message: 'Vous n\'avez pas le droit de faire celà' })
+        // }
         const filename = post.urlImage.split('/images/')[1];
         console.log(filename);
         fs.unlink(`images/${filename}`, () => {
-            Post.destroy({ where: { id : id }, raw: true})
-            .then(post => {
-                if (post === 0) return res.status(404).json({ message: 'post innexistant' })
-                res.status(200).json({ message: 'post supprimé.' })
-            })
-            .catch(error => res.status(500).json(error))
+            Post.destroy({ where: { id: id }, raw: true })
+                .then(post => {
+                    if (post === 0) return res.status(404).json({ message: 'post innexistant' })
+                    res.status(200).json({ message: 'post supprimé.' })
+                })
+                .catch(error => res.status(500).json(error))
         });
 
     } catch {
-        return res.status(500).json({ message: 'Erreur de base de donnée'})
+        return res.status(500).json({ message: 'Erreur de base de donnée' })
     }
-    
+
 }
 
 const likeUnlike = async (req, res, next) => {
@@ -108,30 +125,33 @@ const likeUnlike = async (req, res, next) => {
     const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
     const userId = decodedToken.id
     // console.log(like);
-    
-    try{
-        const post = await Post.findOne({where: { id : id }})
 
-        const postLiked = 
-                    { ...post,
-                        userLiked: [] };
+    try {
+        const post = await Post.findOne({ where: { id: id } })
 
-                    switch(like) {
-                        case 0 :
-                            if(postLiked.userLiked.includes(userId)) {
-                                let indexUser = postLiked.userLiked.indexOf(userId)
-                                postLiked.userLiked.splice(indexUser, 1)
-                            }
-                            break;
-                        case 1 : postLiked.userLiked.push(userId)
-                        break;
-                    }
-                   
-                    await Post.update( postLiked, { where: { id : id}})
-                    return res.status(201).json({message: 'Post like/unLike.'})
+        const postLiked =
+        {
+            ...post,
+            userLiked: []
+        };
 
-    }catch{
-        return res.status(500).json({error : "erreur de base de donnée dans le like"})
+        switch (like) {
+            case 0:
+                if (postLiked.userLiked.includes(userId)) {
+                    let indexUser = postLiked.userLiked.indexOf(userId)
+                    postLiked.userLiked.splice(indexUser, 1)
+                }
+                break;
+            case 1: postLiked.userLiked.push(userId)
+                break;
+        }
+
+        await Post.update(postLiked, { where: { id: id } })
+        return res.status(201).json({ message: 'Post like/unLike.' })
+
+    }
+    catch {
+        return res.status(500).json({ error: "erreur de base de donnée dans le like" })
     }
 }
